tests/requirements.txt
pytest

azure-mgmt-resource
azure-identity
msal[broker]==1.24.0b2